refactored lsp
