+++ omnisharp.csproj
@@ -13,6 +13,11 @@
     <Compile Include="deps/metagen.cs" />
     <Compile Include="deps/msgpack/*.cs" />
     <Compile Include="deps/msgpack/Compiler/*.cs" />
+    <Compile Include="src\lsp\connection.cs" />
+    <Compile Include="src\lsp\protocol.cs" />
+    <Compile Include="src\lsp\rpc.cs" />
+    <Compile Include="src\lsp\server.cs" />
+    <Compile Include="src\lsp\services.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!--
