+++ src/lsp/services.cs
@@ -1,0 +1,180 @@
+using System;
+
+namespace bhlsp
+{
+#region -- ABSTRACT --
+  public abstract class BHLSPJsonRpcService
+  {
+  }
+  
+  public abstract class BHLSPGeneralJsonRpcServiceTemplate : BHLSPJsonRpcService
+  {
+    [JsonRpcMethod("initialize")]
+    public abstract RpcResult Initialize(InitializeParams args);
+
+    [JsonRpcMethod("initialized")]
+    public abstract RpcResult Initialized();
+
+    [JsonRpcMethod("shutdown")]
+    public abstract RpcResult Shutdown();
+
+    [JsonRpcMethod("exit")]
+    public abstract RpcResult Exit();
+  }
+
+  public abstract class BHLSPTextDocumentJsonRpcServiceTemplate : BHLSPJsonRpcService
+  {
+    [JsonRpcMethod("textDocument/didOpen")]
+    public abstract RpcResult DidOpenTextDocument(DidOpenTextDocumentParams args);
+
+    [JsonRpcMethod("textDocument/didChange")]
+    public abstract RpcResult DidChangeTextDocument(DidChangeTextDocumentParams args);
+
+    [JsonRpcMethod("textDocument/willSave")]
+    public abstract RpcResult WillSaveTextDocument(WillSaveTextDocumentParams args);
+
+    [JsonRpcMethod("textDocument/willSaveWaitUntil")]
+    public abstract RpcResult WillSaveWaitUntilTextDocument(WillSaveTextDocumentParams args);
+
+    [JsonRpcMethod("textDocument/didSave")]
+    public abstract RpcResult DidSaveTextDocument(DidSaveTextDocumentParams args);
+
+    [JsonRpcMethod("textDocument/didClose")]
+    public abstract RpcResult DidCloseTextDocument(DidCloseTextDocumentParams args);
+
+    [JsonRpcMethod("textDocument/completion")]
+    public abstract RpcResult Completion(CompletionParams args);
+
+    [JsonRpcMethod("completionItem/resolve")]
+    public abstract RpcResult ResolveCompletionItem(CompletionItem args);
+
+    [JsonRpcMethod("textDocument/hover")]
+    public abstract RpcResult Hover(TextDocumentPositionParams args);
+
+    [JsonRpcMethod("textDocument/signatureHelp")]
+    public abstract RpcResult SignatureHelp(TextDocumentPositionParams args);
+
+    [JsonRpcMethod("textDocument/references")]
+    public abstract RpcResult FindReferences(ReferenceParams args);
+
+    [JsonRpcMethod("textDocument/documentHighlight")]
+    public abstract RpcResult DocumentHighlight(TextDocumentPositionParams args);
+
+    [JsonRpcMethod("textDocument/documentSymbol")]
+    public abstract RpcResult DocumentSymbols(DocumentSymbolParams args);
+
+    [JsonRpcMethod("textDocument/documentColor")]
+    public abstract RpcResult DocumentColor(DocumentColorParams args);
+
+    [JsonRpcMethod("textDocument/colorPresentation")]
+    public abstract RpcResult ColorPresentation(ColorPresentationParams args);
+
+    [JsonRpcMethod("textDocument/formatting")]
+    public abstract RpcResult DocumentFormatting(DocumentFormattingParams args);
+
+    [JsonRpcMethod("textDocument/rangeFormatting")]
+    public abstract RpcResult DocumentRangeFormatting(DocumentRangeFormattingParams args);
+
+    [JsonRpcMethod("textDocument/onTypeFormatting")]
+    public abstract RpcResult DocumentOnTypeFormatting(DocumentOnTypeFormattingParams args);
+
+    [JsonRpcMethod("textDocument/definition")]
+    public abstract RpcResult GotoDefinition(TextDocumentPositionParams args);
+
+    [JsonRpcMethod("textDocument/typeDefinition")]
+    public abstract RpcResult GotoTypeDefinition(TextDocumentPositionParams args);
+
+    [JsonRpcMethod("textDocument/implementation")]
+    public abstract RpcResult GotoImplementation(TextDocumentPositionParams args);
+
+    [JsonRpcMethod("textDocument/codeAction")]
+    public abstract RpcResult CodeAction(CodeActionParams args);
+
+    [JsonRpcMethod("textDocument/codeLens")]
+    public abstract RpcResult CodeLens(CodeLensParams args);
+
+    [JsonRpcMethod("codeLens/resolve")]
+    public abstract RpcResult ResolveCodeLens(CodeLens args);
+
+    [JsonRpcMethod("textDocument/documentLink")]
+    public abstract RpcResult DocumentLink(DocumentLinkParams args);
+
+    [JsonRpcMethod("documentLink/resolve")]
+    public abstract RpcResult ResolveDocumentLink(DocumentLink args);
+
+    [JsonRpcMethod("textDocument/rename")]
+    public abstract RpcResult Rename(RenameParams args);
+
+    [JsonRpcMethod("textDocument/foldingRange")]
+    public abstract RpcResult FoldingRange(FoldingRangeParams args);
+  }
+
+  public abstract class BHLSPWorkspaceJsonRpcServiceTemplate : BHLSPJsonRpcService
+  {
+    [JsonRpcMethod("workspace/didChangeWorkspaceFolders")]
+    public abstract RpcResult DidChangeWorkspaceFolders(DidChangeWorkspaceFoldersParams args);
+
+    [JsonRpcMethod("workspace/didChangeConfiguration")]
+    public abstract RpcResult DidChangeConfiguration(DidChangeConfigurationParams args);
+
+    [JsonRpcMethod("workspace/didChangeWatchedFiles")]
+    public abstract RpcResult DidChangeWatchedFiles(DidChangeWatchedFilesParams args);
+
+    [JsonRpcMethod("workspace/symbol")]
+    public abstract RpcResult Symbol(WorkspaceSymbolParams args);
+
+    [JsonRpcMethod("workspace/executeCommand")]
+    public abstract RpcResult ExecuteCommand(ExecuteCommandParams args);
+  }
+#endregion
+  
+  public class BHLSPGeneralJsonRpcService : BHLSPGeneralJsonRpcServiceTemplate
+  {
+    public override RpcResult Initialize(InitializeParams args)
+    {
+      ServerCapabilities capabilities = new ServerCapabilities
+      {
+        textDocumentSync = new TextDocumentSyncOptions
+        {
+          openClose = true,
+          change = TextDocumentSyncKind.Incremental,
+          save = new SaveOptions
+          {
+            includeText = true
+          }
+        },
+        
+        signatureHelpProvider = new SignatureHelpOptions
+        {
+          triggerCharacters = new []{"(", ","}
+        }
+      };
+      
+      return RpcResult.Success(new InitializeResult
+      {
+        capabilities = capabilities,
+        serverInfo = new InitializeResult.InitializeResultsServerInfo
+        {
+          name = "bhlsp",
+          version = "0.0.1"
+        }
+      });
+    }
+
+    public override RpcResult Initialized()
+    {
+      return RpcResult.Success(null);
+    }
+
+    public override RpcResult Shutdown()
+    {
+      return RpcResult.Success(null);
+    }
+
+    public override RpcResult Exit()
+    {
+      Environment.Exit(0);
+      return RpcResult.Success(null);
+    }
+  }
+}
